Change Enum to int in SaitekPanelLEDPosition

+++ src/DCSFlightpanels/PanelUserControls/BackLitPanelUserControl.xaml.cs
@@ -213,7 +213,7 @@
                 var imageName = contextMenu.Tag.ToString();
                 var position = BacklitPanelBIP.GetLedPosition(imageName);
 
-                var ledConfigsWindow = new LEDConfigsWindow("Set configuration for LED : " + position, new SaitekPanelLEDPosition(position), _backlitPanelBIP.GetLedDcsBiosOutputs(position), _backlitPanelBIP);
+                var ledConfigsWindow = new LEDConfigsWindow("Set configuration for LED : " + position, new SaitekPanelLEDPosition((int)position), _backlitPanelBIP.GetLedDcsBiosOutputs(position), _backlitPanelBIP);
                 if (ledConfigsWindow.ShowDialog() == true)
                 {
                     //must include position because if user has deleted all entries then there is nothing to go after regarding position
